fix sign of loopback edge
The typical mtk models will have terminal on the outside,
i.e. they "inject" into the terminal, the network "injects" back
and thus a "positive" injection in to the network is recied as a
negative value on the output. This is consistent with how we
handle it on defualt nodes but still confusing and needs proper
documentation